Blue 10004 program awards (#163)
* Add uploading, create check, and delete check

BLUE-10004

* update postman tests

* Update config.py

added TODO for future logging implementation

* Update config.py

---------

+++ app/actions/programs/awards/program_award_actions.py
@@ -1,5 +1,9 @@
+from app.exceptions import ExceptionHandling
 from app.actions.base_actions import BaseActions
+from app.actions.upload import UploadActions
 from app.models.programs.program_award_models import ProgramAwardModelDB, ProgramAwardUpdate
+from app.models.segments import SegmentAward
+
 
 
 class ProgramAwardActions:
@@ -26,9 +30,24 @@ class ProgramAwardActions:
 			],
 		)
 
-	@staticmethod
-	async def create_award(path_params: dict, award_obj):
+	@classmethod
+	async def get_upload_url(cls, path_params: dict, file_name: str, upload_type: str):
+		if upload_type == "image":
+			return await UploadActions.generate_upload_url(
+				upload_type,
+				file_name,
+				path_params.get("client_uuid"),
+				path_params.get("program_award_9char"),
+				path_params.get("program_9char")
+			)
+
+		raise ExceptionHandling.custom400(f"Upload type {upload_type} is not supported")
+
+	@classmethod
+	async def create_award(cls,path_params: dict, award_obj):
 		if isinstance(award_obj, list):
+			to_create = []
+			return_list = []
 			award_models = [
 				ProgramAwardModelDB(
 					**award.dict(),
@@ -37,21 +56,40 @@ class ProgramAwardActions:
 					client_award_9char=path_params["client_award_9char"]
 				) for award in award_obj
 			]
-		else:
-			award_models = ProgramAwardModelDB(
-				**award_obj.dict(),
-				client_uuid=path_params["client_uuid"],
-				program_9char=path_params["program_9char"],
-				client_award_9char=path_params["client_award_9char"]
-			)
-		return await BaseActions.create(award_models)
+			for award in award_models:
+				existing_award = await cls.check_if_award_exists(award.uuid)
+				if existing_award:
+					return_list.append(existing_award)
+				else:
+					to_create.append(award)
+			if to_create:
+				return_list.extend(await BaseActions.create(to_create))
+			return return_list
+
+		award_model = ProgramAwardModelDB(
+			**award_obj.dict(),
+			client_uuid=path_params["client_uuid"],
+			program_9char=path_params["program_9char"],
+			client_award_9char=path_params["client_award_9char"]
+		)
+		existing_award = await cls.check_if_award_exists(award_model.uuid)
+		if existing_award:
+			return existing_award
+
+		return await BaseActions.create(award_model)
 
 
-	@staticmethod
+	@classmethod
 	async def update_award(
+		cls,
 		path_params: dict,
 		award_updates: ProgramAwardUpdate
 	):
+		if award_updates.name:
+			await cls.check_if_award_exists_by_name(path_params.get("client_uuid"), award_updates.name)
+		if award_updates.hero_image:
+			award_updates.hero_image, _ = await UploadActions.verify_upload_file("image", award_updates.hero_image)
+			# TODO: add s3 query to check if file exists and is valid
 		return await BaseActions.update(
 			ProgramAwardModelDB,
 			[
@@ -64,6 +102,19 @@ class ProgramAwardActions:
 
 	@staticmethod
 	async def delete_award(path_params: dict):
+		segment =  await BaseActions.get_one_where(
+			SegmentAward,
+			[
+				SegmentAward.client_uuid == path_params["client_uuid"],
+				SegmentAward.program_9char == path_params["program_9char"],
+				SegmentAward.program_award_9char == path_params["program_award_9char"]
+			],
+			False
+		)
+		if segment:
+			return await ExceptionHandling.custom409(
+				f"Cannot delete program award {path_params.program_award_9char} is currently in use by segment: {segment.name}."
+			)
 		return await BaseActions.delete_one(
 			ProgramAwardModelDB,
 			[
@@ -72,3 +123,29 @@ class ProgramAwardActions:
 				ProgramAwardModelDB.program_award_9char == path_params["program_award_9char"]
 			],
 		)
+
+	@staticmethod
+	async def check_if_award_exists_by_name(client_uuid: str, name: str, error=True):
+		award = await BaseActions.check_if_exists(
+			ProgramAwardModelDB,
+			[
+				ProgramAwardModelDB.name == name,
+				ProgramAwardModelDB.client_uuid == client_uuid
+			]
+		)
+		if award and error:
+			return await ExceptionHandling.custom409(f"Client award with name '{name}' already exists.")
+		elif award and not error:
+			return award
+		else:
+			return None
+
+	@staticmethod
+	async def check_if_award_exists(award_uuid: str):
+		award = await BaseActions.check_if_exists(
+			ProgramAwardModelDB,
+			[
+				ProgramAwardModelDB.uuid == award_uuid
+			]
+		)
+		return award
